add some special casing for st
St doesn't support drag and motion versions of mouse events; it also does not
support the ATTR_DIM - avoid using these with st for now to keep st from
complaining. The behavior of st without these kludges is still right though;
so perhaps the right thing to do is to expect st to not complain about
unhandled control codes?